edit: move wordpress plugins to misc
