+++ extractor/filesystem/list/list.go
@@ -59,8 +59,8 @@ import (
 	"github.com/google/osv-scalibr/extractor/filesystem/language/rust/cargotoml"
 	"github.com/google/osv-scalibr/extractor/filesystem/language/swift/packageresolved"
 	"github.com/google/osv-scalibr/extractor/filesystem/language/swift/podfilelock"
-	"github.com/google/osv-scalibr/extractor/filesystem/language/wordpress/plugins"
 	"github.com/google/osv-scalibr/extractor/filesystem/misc/vscodeextensions"
+	wordpressplugins "github.com/google/osv-scalibr/extractor/filesystem/misc/wordpress/plugins"
 	"github.com/google/osv-scalibr/extractor/filesystem/os/apk"
 	"github.com/google/osv-scalibr/extractor/filesystem/os/cos"
 	"github.com/google/osv-scalibr/extractor/filesystem/os/dpkg"
@@ -163,7 +163,6 @@ var (
 	// PHP extractors.
 	PHP = InitMap{composerlock.Name: {composerlock.New}}
 	// Swift extractors.
-
 	Swift = InitMap{
 		packageresolved.Name: {packageresolved.NewDefault},
 		podfilelock.Name:     {podfilelock.NewDefault},
@@ -171,9 +170,6 @@ var (
 
 	// Containers extractors.
 	Containers = InitMap{containerd.Name: {containerd.NewDefault}} // Wordpress extractors.
-
-	// Wordpress extractors.
-	Wordpress = InitMap{plugins.Name: {plugins.NewDefault}}
 
 	// OS extractors.
 	OS = InitMap{
@@ -192,8 +188,10 @@ var (
 		macapps.Name:  {macapps.NewDefault},
 	}
 
+	// Misc extractors.
 	Misc = InitMap{
 		vscodeextensions.Name: {vscodeextensions.New},
+		wordpressplugins.Name: {wordpressplugins.NewDefault},
 	}
 
 	// Collections of extractors.
@@ -221,7 +219,6 @@ var (
 		OS,
 		Misc,
 		Containers,
-		Wordpress,
 	)
 
 	extractorNames = concat(All, InitMap{
@@ -245,7 +242,7 @@ var (
 		"sbom":       vals(SBOM),
 		"os":         vals(OS),
 		"containers": vals(Containers),
-		"wordpress":  vals(Wordpress),
+		"misc":       vals(Misc),
 
 		// Collections.
 		"default": vals(Default),
