+++ extractor/filesystem/misc/wordpress/plugins/plugins.go
