+++ extractor/filesystem/misc/wordpress/plugins/testdata/valid
